src/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
click==8.1.8
colorama==0.4.6
config==0.5.1
coverage==7.6.10
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
pydantic==2.10.6
pydantic_core==2.27.2
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==14.2